remove AbstractDynamicTableInfo
DocTableInfo was the only class that extended from it.
